% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_project_gant_chart.R
\name{clean_projects}
\alias{clean_projects}
\title{Helper function: Clean projects}
\usage{
clean_projects(projects_json)
}
\arguments{
\item{projects_json}{projects_json (default: "https://kwb-r.github.io/kwb.site/projects.json")}
}
\value{
tibble with title and extracted project "id"
}
\description{
Helper function: Clean projects
}
