% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project.R
\name{extract_funder_ids}
\alias{extract_funder_ids}
\title{Helper function: extract funder ids}
\usage{
extract_funder_ids(funder_logo_urls)
}
\arguments{
\item{funder_logo_urls}{vector with funder urls}
}
\value{
vector with funder names
}
\description{
Helper function: extract funder ids
}
