% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_project_gant_chart.R
\name{extract_project_ids}
\alias{extract_project_ids}
\title{Helper function: Extract project IDs}
\usage{
extract_project_ids(title)
}
\arguments{
\item{title}{project title (as retrieved by get_projects() function)}
}
\value{
tibble with title and extracted project "id"
}
\description{
Helper function: Extract project IDs
}
