% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team.R
\name{get_person}
\alias{get_person}
\title{Team: get person}
\usage{
get_person(x)
}
\arguments{
\item{x}{one element of css-class="li-employee-item" (as retrieved by
kwb.site::get_team_url())}
}
\value{
tibble with information of one person
}
\description{
Team: get person
}
\examples{
#' @examples
 \dontrun{
 rs <- kwb.site::rs_connect()
 team_html <- kwb.site::get_team_html(rs = rs)
 team <- team_html \%>\%
 rvest::html_nodes("li.employee-item")
 kwb.site::get_person(team[1])
 }
}
