% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_urls.R
\name{get_project_urls}
\alias{get_project_urls}
\title{Get Project URLs}
\usage{
get_project_urls(language = "en", debug = TRUE)
}
\arguments{
\item{language}{website language "en" (English) or "de" (German), default: "en"}

\item{debug}{print debug messages (default: TRUE)}
}
\value{
character vector with project urls
}
\description{
Get Project URLs
}
\examples{
project_urls_en <- get_project_urls("en")
head(project_urls_en)

\dontrun{project_urls_de <- get_project_urls("de")
head(project_urls_de)
}

}
