% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects.R
\name{get_projects}
\alias{get_projects}
\title{Get Projects Info (English and German)}
\usage{
get_projects(debug = TRUE)
}
\arguments{
\item{debug}{print debug messages (default: TRUE)}
}
\value{
tibble with infos for all projects in English and German
}
\description{
Get Projects Info (English and German)
}
\examples{
\dontrun{
project_infos <- get_projects()
project_infos
}
}
