% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team.R
\name{get_team}
\alias{get_team}
\title{Team: get all team members}
\usage{
get_team(url = "https://www.kompetenz-wasser.de/de/ueber-uns/team-2/", rs)
}
\arguments{
\item{url}{url to team site (default:  "https://www.kompetenz-wasser.de/de/ueber-uns/team-2/")}

\item{rs}{connection to RSelenium (kwb.site::rs_connect()))}
}
\value{
tibble with information of all persons
}
\description{
Team: get all team members
}
\examples{
#' @examples
 \dontrun{
 rs <- kwb.site::rs_connect()
 kwb.site::get_team(rs = rs)
 }
}
