% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team.R
\name{get_team_html}
\alias{get_team_html}
\title{Get Team HTML}
\usage{
get_team_html(url = "https://www.kompetenz-wasser.de/de/ueber-uns/team-2/", rs)
}
\arguments{
\item{url}{url to team site (default:  "https://www.kompetenz-wasser.de/de/ueber-uns/team-2/")}

\item{rs}{connection to RSelenium (kwb.site::rs_connect())}
}
\value{
html source code of team website (after clicking "load-more" and
"open-details" buttons)
}
\description{
Get Team HTML
}
\examples{
 \dontrun{
 rs <- rs_connect(browser = "firefox", geckover = "0.25.0")
 team_html <- kwb.site::get_team_html(rs = rs)
 team_html
 }
}
