% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_project_gant_chart.R
\name{plot_gantt_chart_project}
\alias{plot_gantt_chart_project}
\title{Create KWB Projects Gant Chart}
\usage{
plot_gantt_chart_project(
  projects_json = "https://kwb-r.github.io/kwb.site/projects.json",
  tag_selection = "department",
  language_selection = "en",
  interactive = FALSE,
  interactive_export = FALSE,
  interactive_export_dir = ".",
  alpha = 0.5
)
}
\arguments{
\item{projects_json}{default: "https://kwb-r.github.io/kwb.site/projects.json")}

\item{tag_selection}{default: "department" (or "topic")}

\item{language_selection}{default: "en" (or "de")}

\item{interactive}{should interactive ploty be used (default: FALSE)}

\item{interactive_export}{should interactive plotly graph exported?
default: FALSE}

\item{interactive_export_dir}{export directory in case of interactive = TRUE,
default: "."}

\item{alpha}{alpha (default: 0.5)}
}
\value{
KWB projects gant chart
}
\description{
Create KWB Projects Gant Chart
}
\examples{
\dontrun{
projects_by_department_en <- plot_project_gant_chart()
plot_project_gant_chart(interactive = TRUE)
projects_by_department_de <- plot_project_gant_chart(language_selection = "de")
projects_by_topic_en <- plot_project_gant_chart(tag_selection = "topic")
projects_by_topic_de <- plot_project_gant_chart(tag_selection = "topic", language_selection = "de")
}
}
