% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_project_gant_chart.R
\name{plot_gantt_chart_project_by_funder}
\alias{plot_gantt_chart_project_by_funder}
\title{Create KWB Projects by Funder Gantt Chart}
\usage{
plot_gantt_chart_project_by_funder(
  projects_json = "https://kwb-r.github.io/kwb.site/projects.json",
  language_selection = "en",
  interactive = FALSE,
  interactive_export = FALSE,
  interactive_export_dir = ".",
  alpha = 0.5
)
}
\arguments{
\item{projects_json}{default: "https://kwb-r.github.io/kwb.site/projects.json")}

\item{language_selection}{default: "en" (or "de")}

\item{interactive}{should interactive ploty be used (default: FALSE)}

\item{interactive_export}{should interactive plotly graph exported?
default: FALSE}

\item{interactive_export_dir}{export directory in case of interactive = TRUE,
default: "."}

\item{alpha}{alpha (default: 0.5)}
}
\value{
KWB projects by funder gantt chart
}
\description{
Create KWB Projects by Funder Gantt Chart
}
